// gzoom2-fe/app/src/app/api/model/glAccount.ts
import { CustomMethod } from "./customMethod";


export class GlAccount {
    glAccountId: string;
    glAccountTypeId: string;
    glAccountClassId?: string;
    glResourceTypeId?: string;
    glXbrlClassId?: string;
    parentGlAccountId?: string;
    accountCode: string;
    accountName: string;
    description?: string;
    productId?: string;
    externalId?: string;
    postedBalance?: string;
    lastUpdatedStamp: string;
    lastUpdatedTxStamp: string;
    createdStamp: string;
    createdTxStamp: string;
    debitCreditDefault: string;
    defaultUomId: string;
    periodicalAbsoluteEnumId: string;
    referencedAccountId?: string;
    accountTypeEnumId: string;
    localNameContentId?: string;
    childFolderFile?: string;
    hasCompetence?: string;
    detectOrgUnitIdFlag: string;
    currentStatusId: string;
    inputEnumId: string;
    periodTypeId: string;
    fromDate?: Date;
    thruDate?: Date;
    dataSourceId?: string;
    detailEnumId: string;
    calcCustomMethodId?: string;
    prioCalc?: number;
    weMeasureTypeEnumId: string;
    weScoreRangeEnumId: string;
    weScoreConvEnumId: string;
    weAlertRuleEnumId: string;
    uomRangeId?: string;
    weWithoutPerf: string;
    source?: string;
    trendEnumId: string;
    respCenterId?: string;
    respCenterRoleTypeId?: string;
    emplPositionTypeId?: string;
    accountNameLang?: string;
    descriptionLang?: string;
    sourceLang?: string;
    roleTypeId?: string;
    sequenceId: number;
    targetPeriodEnumId: string;
    weWithoutTarget: string;
    lastModifiedByUserLogin: string;
    createdByUserLogin: string;
}